+++ test_opengl/Source/Test_Mix.cpp
@@ -4,7 +4,7 @@
 #include "Glwindow.h"
 #include "GlProgram.h"
 #include "Camera.h"
-#include "Math.hpp"
+#include "YQMath.h"
 
 #include "glad/glad.h"
 #include "GLFW/glfw3.h"
@@ -172,8 +172,8 @@ void Test::Test_Mix()
 
 	std::function<void(double, double)> mouseMove = [&](double xpos, double ypos) {
 		camera->OnMoveView((float)xpos - lastX, (float)ypos - lastY);
-		lastX = xpos;
-		lastY = ypos;
+		lastX = (float)xpos;
+		lastY = (float)ypos;
 	};
 
 	camera->SetCameraPos({ 1.7f ,1.5f ,5.0f });
