+++ test_opengl/Source/Test_TemplateTesting.cpp
@@ -4,7 +4,7 @@
 #include "GlWindow.h"
 #include "GlProgram.h"
 #include "Camera.h"
-#include "Math.hpp"
+#include "YQMath.h"
 
 #include "glad/glad.h"
 #include "GLFW/glfw3.h"
@@ -144,9 +144,9 @@ void Test::Test_TemplateTesting()
 	float yaw = 0, pitch = 0;
 
 	std::function<void(double, double)> mouseMove = [&](double xpos, double ypos) {
-		camera.OnMoveView(xpos - lastX, ypos - lastY);
-		lastX = xpos;
-		lastY = ypos;
+		camera.OnMoveView((float)xpos - lastX, (float)ypos - lastY);
+		lastX = (float)xpos;
+		lastY = (float)ypos;
 	};
 
 	camera.SetCameraPos({ 1.7f ,1.5f ,5.0f });
